Split Ztm2025GeoMapViewer into OpenLayersGeoMap
